pyproject.toml
[project]
name = "pyllmvalidate-cli"
version = "0.1.0"
description = "A CLI tool for validating code patches using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.44",
    "openai>=1.82.1",
    "pydantic>=2.11.5",
]

[project.scripts]
pyllmvalidate = "src.pyllmvalidate.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "examples/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
